Begun major rewrite of the Tessen Module, going to separate into two files.
Generate: This will generate fan curves for each unique trading period id.
That is the general incremental stacks mapping energy and reserve offers
together.

Visualise: This will take the pregenerated data and apply filters to it in
order to construct the fan curve. This should be a loose coupling. It should
not do any major processing itself, just filter and plot the data although
some sorting will be necessary.

Goal is to completely separate out the processing and plotting.
In this way a JavaScript version could probably be completed using D3 to
quickly flick between different fans