Fixed the legend issue, turns out it was an issue with the reserve line calculation
I fucked up the signs >= and > when calculating the low price tranch.
This led to a lot of overlapping which as alpha values were non unity
was generally leading to weird colour permutations and wrecking my day

Fuck that shit. I've fixed this issue by adding an initial zero check, the
zero check is necessary as behaviour for the very first tranche is
slightly different than for subsequent tranches.

Without the check any energy offers priced at zero would be ignored.